sql: move `vars` module into new `session` module
To make space for a forthcoming `user` module.
