+++ src/adapter/src/notice.rs
@@ -14,7 +14,8 @@ use chrono::{DateTime, Utc};
 use mz_controller::clusters::ClusterStatus;
 use mz_ore::str::StrExt;
 use mz_repr::strconv;
-use mz_sql::{ast::NoticeSeverity, vars::IsolationLevel};
+use mz_sql::ast::NoticeSeverity;
+use mz_sql::session::vars::IsolationLevel;
 
 /// Notices that can occur in the adapter layer.
 ///
