+++ src/sql/src/session/vars.rs
